Moved ImageCommentsViewController to production
