+++ EssentialApp/EssentialAppTests/ImageCommentsUIIntegrationTests.swift
@@ -10,71 +10,6 @@ import XCTest
 import EssentialFeed
 import EssentialFeediOS
 
-class ImageCommentsViewController : UITableViewController, ImageCommentView, ImageCommentLoadingView, ImageCommentErrorView {
-	
-	public var presenter: ImageCommentPresenter?
-	public var loader: ImageCommentLoader?
-	
-	public let errorView = UILabel()
-	
-	private var tableModel = [ImageCommentViewModel]() {
-		didSet { tableView.reloadData() }
-	}
-	
-	override func viewDidLoad() {
-		super.viewDidLoad()
-		
-		refreshControl = UIRefreshControl()
-		refreshControl?.addTarget(self, action: #selector(refresh), for: .valueChanged)
-		
-		tableView.register(ImageCommentCell.self, forCellReuseIdentifier: "\(ImageCommentCell.self)")
-		
-		refresh()
-	}
-	
-	@objc public func refresh() {
-		presenter?.didStartLoadingComments()
-		loader?.load { [weak self] result in
-			switch result {
-			case let .success(imageComments):
-				self?.presenter?.didFinishLoadingComments(with: imageComments)
-				break
-			case let .failure(error):
-				self?.presenter?.didFinishLoadingComments(with: error)
-				break
-			}
-		}
-	}
-	
-	func display(_ viewModel: ImageCommentsViewModel) {
-		tableModel = viewModel.comments
-	}
-	
-	func display(_ viewModel: ImageCommentLoadingViewModel) {
-		if viewModel.isLoading {
-			refreshControl?.beginRefreshing()
-		} else {
-			refreshControl?.endRefreshing()
-		}
-	}
-	
-	func display(_ viewModel: ImageCommentErrorViewModel) {
-		errorView.text = viewModel.message
-	}
-	
-	override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-		return tableModel.count
-	}
-	
-	override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-		let model = tableModel[indexPath.row]
-		
-		let cell = tableView.dequeueReusableCell(withIdentifier: "\(ImageCommentCell.self)") as! ImageCommentCell
-		cell.configure(model)
-		return cell
-	}
-}
-
 class ImageCommentsUIComposer {
 	
 	static func imageCommentsComposedWith(loader: ImageCommentLoader) -> ImageCommentsViewController {
